requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
click==8.1.3
django-environ==0.9.0
django-filter==22.1
Django==4.1.4
djangorestframework==3.14.0
kombu==5.2.4
pip==22.3.1
prompt-toolkit==3.0.36
pytz==2022.7
redis==4.4.0
setuptools==65.6.3
six==1.16.0
sqlparse==0.4.3
vine==5.0.0
wcwidth==0.2.5
wheel==0.38.4
psycopg2-binary==2.9.5

